+++ TOReachabilityExample (iOS)/CollectionViewCell.h
@@ -12,8 +12,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface CollectionViewCell : UICollectionViewCell
 
-@property (weak, nonatomic) IBOutlet UIImageView *backgroundImageView;
-@property (weak, nonatomic) IBOutlet UIImageView *highlightedImageView;
+@property (weak, nonatomic) IBOutlet UIView *unhighlightedView;
+@property (weak, nonatomic) IBOutlet UIView *highlightedView;
 @property (weak, nonatomic) IBOutlet UIImageView *imageView;
 @property (weak, nonatomic) IBOutlet UILabel *titleLabel;