+++ CHANGELOG.md
@@ -10,6 +10,21 @@ x.y.z Release Notes (yyyy-MM-dd)
 ### Compatibility
 * None.
 
+1.0.0 Release Notes (2023-05-13)
+=============================================================
+
+### Enhancements
+* For Mac and Apple TV use cases, `TOReachabilityStatus` has been reworded from `WiFi` to `Available` to make more sense on devices that may also use Ethernet for their network connections.
+
+### Fixed
+* Fixed nullability annotation warnings.
+
+### Compatibility
+* Bumped minimum version to iOS 11.0.
+* Added support for tvOS 11.0 and macOS 10.13.
+
+* Initial Release! 🎉
+
 0.1.0 Release Notes (2018-03-15)
 =============================================================